// src/model/indemnizacion.ts
export class Indemnizacion {
  constructor() {
    this.idCliente = 0;
    this.nombresCliente = "";
    this.telefonoCliente = "";
    this.direccionCliente = "";
    this.tipo = "";
    this.valorRenta = 0;
    this.meses = 0;
    this.interes = 0;
    this.valorTotal = 0;
  }



  idCliente: number;
  nombresCliente: string;
  telefonoCliente: string;
  direccionCliente: string;
  tipo: string;
  valorRenta: number;
  meses: number;
  interes: number;
  valorTotal: number;
}